Replace old CellImg with revised version and fix resulting errors
