+++ src/main/java/net/imglib2/img/cell/AbstractCellImg.java
@@ -11,13 +11,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -34,52 +34,43 @@
 
 package net.imglib2.img.cell;
 
+import net.imglib2.IterableInterval;
+import net.imglib2.RandomAccessible;
 import net.imglib2.img.AbstractNativeImg;
 import net.imglib2.img.Img;
 import net.imglib2.type.NativeType;
+import net.imglib2.util.Fraction;
 
 /**
  * Abstract superclass for {@link Img} types that divide their underlying data
  * into cells.
- * 
+ *
  * @author Mark Hiner
  * @author Tobias Pietzsch
  */
-public abstract class AbstractCellImg< T extends NativeType< T >, A, C extends AbstractCell< A >, F extends AbstractCellImgFactory< T > > extends AbstractNativeImg< T, A >
+public abstract class AbstractCellImg<
+				T extends NativeType< T >,
+				A,
+				C extends Cell< A >,
+				I extends RandomAccessible< C > & IterableInterval< C > >
+		extends AbstractNativeImg< T, A >
 {
+	protected final CellGrid grid;
 
-	final protected F factory;
+	protected final I cells;
 
-	final protected Cells< A, C > cells;
-
-	/**
-	 * Dimensions of a standard cell. Cells on the max border of the image may
-	 * be cut off and have different dimensions.
-	 */
-	final int[] cellDims;
-
-	private static long[] getDimensionsFromCells( final Cells< ?, ? > cells )
+	public AbstractCellImg( final CellGrid grid, final I imgOfCells, final Fraction entitiesPerPixel )
 	{
-		final long[] dim = new long[ cells.numDimensions() ];
-		cells.dimensions( dim );
-		return dim;
-	}
-
-	public AbstractCellImg( final F factory, final Cells< A, C > cells )
-	{
-		super( getDimensionsFromCells( cells ), cells.getEntitiesPerPixel() );
-
-		this.factory = factory;
-		this.cells = cells;
-		cellDims = new int[ cells.numDimensions() ];
-		cells.cellDimensions( cellDims );
+		super( grid.dimensions, entitiesPerPixel );
+		this.grid = grid;
+		this.cells = imgOfCells;
 	}
 
 	/**
-	 * This interface is implemented by all samplers on the {@link CellImg}. It
-	 * allows the container to ask for the cell the sampler is currently in.
+	 * This interface is implemented by all samplers on the {@link AbstractCellImg}. It
+	 * allows to ask for the cell the sampler is currently in.
 	 */
-	public interface CellContainerSampler< T extends NativeType< T >, A, C extends AbstractCell< A > >
+	public interface CellImgSampler< C >
 	{
 		/**
 		 * @return the cell the sampler is currently in.
@@ -91,45 +82,26 @@ public abstract class AbstractCellImg< T extends NativeType< T >, A, C extends A
 	@SuppressWarnings( "unchecked" )
 	public A update( final Object cursor )
 	{
-		return ( ( CellContainerSampler< T, A, C > ) cursor ).getCell().getData();
-	}
-
-	/**
-	 * Get the position of the cell containing the element at {@link position}.
-	 * 
-	 * @param position
-	 *            position of an element in the {@link CellImg}.
-	 * @param cellPos
-	 *            position within cell grid of the cell containing the element.
-	 */
-	protected void getCellPosition( final long[] position, final long[] cellPos )
-	{
-		for ( int d = 0; d < n; ++d )
-			cellPos[ d ] = position[ d ] / cellDims[ d ];
+		// directly get data?
+		return ( ( CellImgSampler< C > ) cursor ).getCell().getData();
 	}
 
 	@Override
-	public CellCursor< T, A, C > cursor()
+	public CellCursor< T, C > cursor()
 	{
 		return new CellCursor<>( this );
 	}
 
 	@Override
-	public CellLocalizingCursor< T, A, C > localizingCursor()
+	public CellLocalizingCursor< T, C > localizingCursor()
 	{
 		return new CellLocalizingCursor<>( this );
 	}
 
 	@Override
-	public CellRandomAccess< T, A, C > randomAccess()
+	public CellRandomAccess< T, C > randomAccess()
 	{
 		return new CellRandomAccess<>( this );
-	}
-
-	@Override
-	public F factory()
-	{
-		return factory;
 	}
 
 	@Override
@@ -139,20 +111,33 @@ public abstract class AbstractCellImg< T extends NativeType< T >, A, C extends A
 	}
 
 	/**
-	 * @return a reference to the {@link Cells} interface which itself gives
-	 *         access to the individual {@link AbstractCell}s through Cursors
-	 *         and RandomAccesses that can return the actual underlying data
-	 *         object.
+	 * Get the underlying image of cells which gives access to the individual
+	 * {@link Cell}s through Cursors and RandomAccesses.
+	 *
+	 * @return the image of cells.
 	 */
-	public Cells< A, C > getCells()
+	public I getCells()
 	{
 		return cells;
 	}
 
+	/**
+	 * Get the {@link CellGrid} which describes the layout of the
+	 * {@link AbstractCellImg}. The grid provides the dimensions of the image, the
+	 * number of cells in each dimension, and the dimensions of individual
+	 * cells.
+	 *
+	 * @return the cell grid layout.
+	 */
+	public CellGrid getCellGrid()
+	{
+		return grid;
+	}
+
 	protected void copyDataTo( final AbstractCellImg< T, ?, ?, ? > copy )
 	{
-		final CellCursor< T, A, C > source = this.cursor();
-		final CellCursor< T, ?, ? > target = copy.cursor();
+		final CellCursor< T, C > source = this.cursor();
+		final CellCursor< T, ? > target = copy.cursor();
 
 		while ( source.hasNext() )
 			target.next().set( source.next() );
