--- src/main/java/net/imglib2/img/cell/AbstractCellImgFactory.java
@@ -1,129 +1,0 @@
-/*
- * #%L
- * ImgLib2: a general-purpose, multidimensional image processing library.
- * %%
- * Copyright (C) 2009 - 2016 Tobias Pietzsch, Stephan Preibisch, Stephan Saalfeld,
- * John Bogovic, Albert Cardona, Barry DeZonia, Christian Dietz, Jan Funke,
- * Aivar Grislis, Jonathan Hale, Grant Harris, Stefan Helfrich, Mark Hiner,
- * Martin Horn, Steffen Jaensch, Lee Kamentsky, Larry Lindsey, Melissa Linkert,
- * Mark Longair, Brian Northan, Nick Perry, Curtis Rueden, Johannes Schindelin,
- * Jean-Yves Tinevez and Michael Zinsmaier.
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * 
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-
-package net.imglib2.img.cell;
-
-import net.imglib2.img.NativeImgFactory;
-import net.imglib2.type.NativeType;
-
-/**
- * Abstract superclass for {@link NativeImgFactory> implementations using
- * 
- * @link CellImg}s.
- * 
- * @author Mark Hiner
- * @author Tobias Pietzsch
- */
-public abstract class AbstractCellImgFactory< T extends NativeType< T > > extends NativeImgFactory< T >
-{
-	protected int[] defaultCellDimensions = { 10 };
-
-	public AbstractCellImgFactory()
-	{}
-
-	public AbstractCellImgFactory( final int cellSize )
-	{
-		defaultCellDimensions[ 0 ] = cellSize;
-	}
-
-	public AbstractCellImgFactory( final int[] cellDimensions )
-	{
-		if ( cellDimensions == null || cellDimensions.length == 0 )
-		{
-			System.err.println( "CellContainerFactory(): cellSize is null. Using equal cell size of " + defaultCellDimensions[ 0 ] );
-			return;
-		}
-
-		for ( int i = 0; i < cellDimensions.length; i++ )
-		{
-			if ( cellDimensions[ i ] <= 0 )
-			{
-				System.err.println( "CellContainerFactory(): cell size in dimension " + i + " is <= 0, using a size of " + defaultCellDimensions[ 0 ] + "." );
-				cellDimensions[ i ] = defaultCellDimensions[ 0 ];
-			}
-		}
-
-		defaultCellDimensions = cellDimensions;
-	}
-
-	protected long[] checkDimensions( long dimensions[] )
-	{
-		if ( dimensions == null || dimensions.length == 0 )
-		{
-			System.err.println( "CellContainerFactory(): dimensionality is null. Creating a 1D cell with size 1." );
-			dimensions = new long[] { 1 };
-		}
-
-		for ( int i = 0; i < dimensions.length; i++ )
-		{
-			if ( dimensions[ i ] <= 0 )
-			{
-				System.err.println( "CellContainerFactory(): size of dimension " + i + " is <= 0, using a size of 1." );
-				dimensions[ i ] = 1;
-			}
-		}
-
-		return dimensions;
-	}
-
-	protected int[] checkCellSize( int[] cellDimensions, final long[] dimensions )
-	{
-		if ( cellDimensions == null )
-		{
-			cellDimensions = new int[ dimensions.length ];
-			for ( int i = 0; i < cellDimensions.length; i++ )
-				cellDimensions[ i ] = defaultCellDimensions[ ( i < defaultCellDimensions.length ) ? i : 0 ];
-		}
-
-		if ( cellDimensions.length != dimensions.length )
-		{
-			// System.err.println(
-			// "CellContainerFactory(): dimensionality of image is unequal to dimensionality of cells, adjusting cell dimensionality."
-			// );
-			final int[] cellDimensionsNew = new int[ dimensions.length ];
-
-			for ( int i = 0; i < dimensions.length; i++ )
-			{
-				if ( i < cellDimensions.length )
-					cellDimensionsNew[ i ] = cellDimensions[ i ];
-				else
-					cellDimensionsNew[ i ] = defaultCellDimensions[ ( i < defaultCellDimensions.length ) ? i : 0 ];
-			}
-
-			cellDimensions = cellDimensionsNew;
-		}
-
-		return cellDimensions;
-	}
-}
