+++ src/main/java/net/imglib2/img/cell/CellCursor.java
@@ -11,13 +11,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -39,11 +39,13 @@ import net.imglib2.Cursor;
 import net.imglib2.type.NativeType;
 
 /**
- * {@link Cursor} on a {@link CellImg}.
- * 
+ * {@link Cursor} on a {@link AbstractCellImg}.
+ *
  * @author Tobias Pietzsch
  */
-public class CellCursor< T extends NativeType< T >, A, C extends AbstractCell< A > > extends AbstractCursor< T > implements AbstractCellImg.CellContainerSampler< T, A, C >
+public class CellCursor< T extends NativeType< T >, C extends Cell< ? > >
+		extends AbstractCursor< T >
+		implements AbstractCellImg.CellImgSampler< C >
 {
 	protected final T type;
 
@@ -62,7 +64,7 @@ public class CellCursor< T extends NativeType< T >, A, C extends AbstractCell< A
 	 */
 	protected boolean isNotLastCell;
 
-	protected CellCursor( final CellCursor< T, A, C > cursor )
+	protected CellCursor( final CellCursor< T, C > cursor )
 	{
 		super( cursor.numDimensions() );
 
@@ -76,12 +78,12 @@ public class CellCursor< T extends NativeType< T >, A, C extends AbstractCell< A
 		type.updateIndex( index );
 	}
 
-	public CellCursor( final AbstractCellImg< T, A, C, ? > container )
+	public CellCursor( final AbstractCellImg< T, ?, C, ? > img )
 	{
-		super( container.numDimensions() );
+		super( img.numDimensions() );
 
-		this.type = container.createLinkedType();
-		this.cursorOnCells = container.cells.cursor();
+		this.type = img.createLinkedType();
+		this.cursorOnCells = img.getCells().cursor();
 
 		reset();
 	}
@@ -99,13 +101,13 @@ public class CellCursor< T extends NativeType< T >, A, C extends AbstractCell< A
 	}
 
 	@Override
-	public CellCursor< T, A, C > copy()
+	public CellCursor< T, C > copy()
 	{
-		return new CellCursor< T, A, C >( this );
+		return new CellCursor<>( this );
 	}
 
 	@Override
-	public CellCursor< T, A, C > copyCursor()
+	public CellCursor< T, C > copyCursor()
 	{
 		return copy();
 	}
