+++ src/test/java/net/imglib2/img/cell/CellCursorTest.java
@@ -11,13 +11,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -39,17 +39,17 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.Random;
 
+import org.junit.Before;
+import org.junit.Test;
+
 import net.imglib2.Cursor;
 import net.imglib2.RandomAccess;
 import net.imglib2.type.numeric.integer.IntType;
 import net.imglib2.util.IntervalIndexer;
 
-import org.junit.Before;
-import org.junit.Test;
-
 /**
  * TODO
- * 
+ *
  */
 public class CellCursorTest
 {
@@ -61,7 +61,7 @@ public class CellCursorTest
 
 	long intDataSum;
 
-	CellImg< IntType, ?, ? > intImg;
+	CellImg< IntType, ? > intImg;
 
 	@Before
 	public void createSourceData()
