+++ src/test/java/net/imglib2/view/RandomAccessibleIntervalCursorBenchmark.java
@@ -11,13 +11,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -44,13 +44,12 @@ import net.imglib2.img.array.ArrayImgFactory;
 import net.imglib2.img.basictypeaccess.array.IntArray;
 import net.imglib2.img.cell.CellImg;
 import net.imglib2.img.cell.CellImgFactory;
-import net.imglib2.img.cell.DefaultCell;
 import net.imglib2.type.numeric.integer.IntType;
 import net.imglib2.util.BenchmarkHelper;
 import net.imglib2.util.IntervalIndexer;
 
 /**
- * 
+ *
  * @author Tobias Pietzsch
  */
 public class RandomAccessibleIntervalCursorBenchmark
@@ -65,7 +64,7 @@ public class RandomAccessibleIntervalCursorBenchmark
 
 	ArrayImg< IntType, IntArray > array2;
 
-	CellImg< IntType, IntArray, DefaultCell< IntArray > > cell;
+	CellImg< IntType, IntArray > cell;
 
 	@SuppressWarnings( "unchecked" )
 	public void setUp()
@@ -73,7 +72,7 @@ public class RandomAccessibleIntervalCursorBenchmark
 		dimensions = new long[] { 207, 103, 1021 };
 		array1 = ( ArrayImg< IntType, IntArray > ) new ArrayImgFactory< IntType >().create( dimensions, new IntType() );
 		array2 = ( ArrayImg< IntType, IntArray > ) new ArrayImgFactory< IntType >().create( dimensions, new IntType() );
-		cell = ( CellImg< IntType, IntArray, DefaultCell< IntArray > > ) new CellImgFactory< IntType >().create( dimensions, new IntType() );
+		cell = ( CellImg< IntType, IntArray > ) new CellImgFactory< IntType >().create( dimensions, new IntType() );
 
 		// fill intData with random values
 		numValues = 1;
